requirements.txt
annotated-types==0.7.0
async-timeout==5.0.1
blinker==1.9.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
exceptiongroup==1.3.0
Flask==3.1.1
flask-cors==5.0.1
greenlet==3.2.2
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
pypdf==5.5.0
pytest==8.3.5
python-dotenv==1.1.0
redis==6.1.0
reportlab==4.4.1
requests==2.32.3
SQLAlchemy==2.0.41
tenacity==9.1.2
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
